pyproject.toml
[project]
name = "tws-traderbot"
version = "0.1.0"
description = "Algorithmic trading bot using Interactive Brokers TWS API and Nautilus Trader"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Bruno" }
]

dependencies = [
    # Core Trading Engine
    "nautilus_trader>=1.180.0",
    "nautilus_ibapi>=10.37.0",  # Repackaged ibapi for Nautilus compatibility
    
    # UI Framework
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    
    # Utilities
    "pytz>=2023.0",
    "requests>=2.31.0",
    
    # Database
    "duckdb>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
traderbot = "src.bot.live_runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
